% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{arrow_dataset_to_duck_db}
\alias{arrow_dataset_to_duck_db}
\title{Create a DuckDB from an Arrow Data Source}
\usage{
arrow_dataset_to_duck_db(
  duck_path = NULL,
  table_name = NULL,
  disconnect = T,
  index_variables = NULL,
  is_unique_index = F,
  read_only = FALSE,
  config = list(),
  arrow_source = NULL,
  arrow_schema_file = NULL,
  partitioning = NULL,
  unify_schemas = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Turns an arrow data source into a duckdb.
}
\details{
Check out \href{https://duckdb.org/docs/}{duckdb documentation}
}
