% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_filter}
\alias{arrow_filter}
\title{Filter Arrow File Set}
\usage{
arrow_filter(
  data,
  id_column = "id_contract_analysis",
  regex_features = c("description_obligation", "name_major_program"),
  regex_values = c("PTSD", "GAY", "GENDER STUDIES", "UKRAINE"),
  exact_features = c("code_product_service", "code_research"),
  exact_values = c("1550", "SR3", "ST3"),
  to_upper = TRUE,
  return_arrow = F,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Filter Arrow File Set
}
\examples{
fpds <- sheldon::arrow_fpds()
arrow_filter(data = fpds, id_column = "id_contract_analysis", regex_features = c("description_obligation", "name_major_program"), regex_values = c("PTSD", "GAY", "GENDER STUDIES", "UKRAINE"), exact_features = c("code_product_service", "code_research"), exact_values = c("1550", "SR3", "ST3"), to_upper = TRUE, return_arrow = F, return_message = T)


}
