% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_filter_exact}
\alias{arrow_filter_exact}
\title{Filter arrow features based on in parameters}
\usage{
arrow_filter_exact(
  data,
  id_column = NULL,
  features = NULL,
  values = NULL,
  return_arrow = F,
  summarise_final_data = T,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Filter arrow features based on in parameters
}
\examples{
fpds <- sheldon::arrow_fpds

arrow_filter_exact(fpds, id_column = "id_contract_analysis", features = c("code_research", "code_product_service"), values = c("SR3", "ST3", "1550"))

}
