% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_filter_regex}
\alias{arrow_filter_regex}
\title{Arrow Dataset Regex Filter}
\usage{
arrow_filter_regex(
  data,
  id_column = "id_contract_analysis",
  features = NULL,
  values = NULL,
  slugs = NULL,
  to_upper = TRUE,
  return_arrow = F,
  summarise_final_data = T,
  return_message = T
)
}
\arguments{
\item{summarise_final_data}{}
}
\value{

}
\description{
Arrow Dataset Regex Filter
}
\examples{

fpds <- sheldon::arrow_fpds()
arrow_filter_regex(data = fpds, id_column = "id_contract_analysis", features = c("description_obligation", "name_major_program"), values = c("PTSD", "GAY", "GENDER STUDIES", "UKRAINE"))


}
