% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_open_data}
\alias{arrow_open_data}
\title{Open an Arrow Data Set}
\usage{
arrow_open_data(
  sources = NULL,
  schema_file = NULL,
  schema = NULL,
  schema_name = NULL,
  assign_schema = T,
  to_duck = F,
  partitioning = NULL,
  unify_schemas = NULL,
  format = c("parquet", "arrow", "ipc", "feather", "csv", "tsv", "text"),
  ...
)
}
\arguments{
\item{sources}{vector of locations}

\item{schema_file}{Location of the schema file}

\item{schema_name}{}

\item{format}{options are `parquet` `ipc` `feather` `csv` `tsv`}
}
\value{

}
\description{
Open an Arrow Data Set
}
\examples{
library(asbtools)
arrow_open_data(sources = "Desktop/data/usa_spending/assistance/", schema_file = "Desktop/data/usa_spending/assistance/2021.gz.parquet", schema_name = "assistance")
asbtools::arrow_open_data(sources  = "Desktop/abresler.github.io/r_packages/govtrackR/data/thousand_talents.tsv.gz", format = "csv") \%>\% count(nameSponsor, sort =  T) \%>\% collect()
arrow_open_data(sources  = "Desktop/abresler.github.io/r_packages/govtrackR/data/thousand_talents.tsv.gz", format = "csv", to_duck = T)

}
