% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_remove_columns}
\alias{arrow_remove_columns}
\title{Remove Setof Arrow Columns}
\usage{
arrow_remove_columns(data, remove_columns = NULL)
}
\arguments{
\item{remove_columns}{}
}
\value{

}
\description{
Remove Setof Arrow Columns
}
