% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow_write_data_set}
\alias{arrow_write_data_set}
\title{Write a dataset to file using arrow}
\usage{
arrow_write_data_set(
  data,
  file_path = NULL,
  format = c("parquet", "feather", "arrow", "ipc", "csv"),
  partitioning = dplyr::group_vars(data),
  base_file_name = paste0("part-{i}.", as.character(format)),
  hive_style = TRUE,
  existing_data_behavior = c("overwrite", "error", "delete_matching"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Write a dataset to file using arrow
}
\examples{
library(asbtools)
library(tidyverse)

data <- ggplot2::diamonds \%>\% tbl_arrow()

arrow_write_data_set(data = data, file_path  = "Desktop/arrow_test_write", format = "csv")
arrow_write_data_set(data = data, file_path  = "Desktop/arrow_test_write", format = "feather")
arrow_write_data_set(data = data, file_path  = "Desktop/arrow_test_write", format = "parquet")
arrow_write_data_set(data = as_tibble(data), file_path  = "Desktop/arrow_test_write", format = "parquet")



}
