% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_message}
\alias{cat_message}
\title{Print a message using cat}
\usage{
cat_message(text = NULL)
}
\arguments{
\item{text}{vector text}
}
\value{
invisible
}
\description{
Print a message using cat
}
\examples{
cat_message(text = "Hello World")
}
