% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_db_index}
\alias{create_db_index}
\title{Create Database index statement}
\usage{
create_db_index(
  table_name = NULL,
  data = NULL,
  index_name = NULL,
  is_unique = F,
  variable = NULL
)
}
\arguments{
\item{table_name}{a table name}

\item{index_name}{if not `NULL` an index name}

\item{variable}{variable name}
}
\value{

}
\description{
Create Database index statement
}
