% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_db_variable_indicies}
\alias{create_db_variable_indicies}
\title{Create Variables Indices}
\usage{
create_db_variable_indicies(
  table_name = NULL,
  variables = NULL,
  data = NULL,
  is_unique = F
)
}
\arguments{
\item{table_name}{a table name}

\item{variables}{vector of variables}

\item{data}{if not `NULL` a `tibble` to validate against}
}
\value{

}
\description{
Create Variables Indices
}
