% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_sql_like}
\alias{create_sql_like}
\title{Create a SQL Like Statement}
\usage{
create_sql_like(
  table_name = NULL,
  features = NULL,
  variables = NULL,
  variable_case = "upper",
  select_sql = "*"
)
}
\arguments{
\item{select_sql}{}
}
\value{

}
\description{
Create a SQL Like Statement
}
