% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{create_sql_regexp_matches}
\alias{create_sql_regexp_matches}
\title{Create a REGEX Match SQL Statement}
\usage{
create_sql_regexp_matches(
  table_name = NULL,
  features = NULL,
  variables = NULL,
  variable_case = "upper",
  select_sql = "*"
)
}
\arguments{
\item{select_sql}{}
}
\value{

}
\description{
Create a REGEX Match SQL Statement
}
