% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{db_duck_add_indicies}
\alias{db_duck_add_indicies}
\title{Add Duck DB Indicies}
\usage{
db_duck_add_indicies(con = NULL, db_indicies)
}
\arguments{
\item{db_indicies}{}
}
\value{

}
\description{
Add Duck DB Indicies
}
