% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{db_duck_connect}
\alias{db_duck_connect}
\title{Create a DuckDB connection}
\usage{
db_duck_connect(
  db_path = NULL,
  file_name = NULL,
  read_only = FALSE,
  config = list()
)
}
\arguments{
\item{db_path}{if not `NULL` path to database}

\item{read_only}{Set to TRUE for read-only operation}

\item{config}{Named list with DuckDB configuration flag}
}
\value{

}
\description{
Create a DuckDB connection
}
