% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{db_duck_disconnect}
\alias{db_duck_disconnect}
\title{Disconnect from DuckDB}
\usage{
db_duck_disconnect(con = NULL, shutdown = T)
}
\arguments{
\item{con}{a connection object}

\item{shutdown}{if `TRUE` shutdown server}
}
\value{

}
\description{
Disconnect from DuckDB
}
