% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{db_duck_write_table}
\alias{db_duck_write_table}
\title{Write table to duck db connection}
\usage{
db_duck_write_table(
  con = NULL,
  data = NULL,
  table_name = NULL,
  db_indicies = NULL,
  drop_existing_table = F,
  append_table = TRUE,
  overwrite_table = FALSE,
  row.names = FALSE
)
}
\arguments{
\item{con}{a duckdb connection}

\item{data}{a table name}

\item{table_name}{table table}

\item{db_indicies}{if not `NULL` vector of index statements}

\item{drop_existing_table}{if `TRUE` drop a table that exists}

\item{append_table}{if `TRUE` appends table}

\item{overwrite_table}{if `TRUE` overrites existing table}

\item{row.names}{if `TRUE` rownames}
}
\value{

}
\description{
Write table to duck db connection
}
\examples{

library(asbtools)

}
