% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{db_filter_regexp}
\alias{db_filter_regexp}
\title{Execute regexp filter search on a database}
\usage{
db_filter_regexp(
  con = NULL,
  table_name = NULL,
  select_sql = "*",
  features = NULL,
  variables = NULL,
  variable_case = "upper",
  return_query = T,
  to_arrow_table = F
)
}
\arguments{
\item{to_arrow_table}{}
}
\value{

}
\description{
Execute regexp filter search on a database
}
