% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{db_table_features}
\alias{db_table_features}
\title{Duck Database Features}
\usage{
db_table_features(con, table_name)
}
\arguments{
\item{table_name}{}
}
\value{

}
\description{
Duck Database Features
}
\examples{
db_table_features(con = con, table_name = "contract_solicitations")
}
