% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_item}
\alias{delete_item}
\title{Remove an Item or Folder}
\usage{
delete_item(path = NULL, recursive = T, force = T, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Remove an Item or Folder
}
