% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{filter_across}
\alias{filter_across}
\title{Filter across features}
\usage{
filter_across(
  data,
  filter_across_columns = NULL,
  exclude_filters = NULL,
  keep_filters = NULL
)
}
\arguments{
\item{keep_filters}{}
}
\value{

}
\description{
Filter across features
}
