% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{future_method}
\alias{future_method}
\title{Setup a future method}
\usage{
future_method(method = NULL, remotes = NULL)
}
\arguments{
\item{method}{if not \code{NULL} sets a `future` method options \itemize{
\item sequential: Resolves futures sequentially in the current R process.
\item transparent: Resolves futures sequentially in the current R process and assignments will be done to the calling environment. Early stopping is enabled by default.
\item multisession: Resolves futures asynchronously (in parallel) in separate R sessions running in the background on the same machine.
\item multicore: Resolves futures asynchronously (in parallel) in separate forked R processes running in the background on the same machine. Not supported on Windows.
\item multiprocess: If multicore evaluation is supported, that will be used, otherwise multisession evaluation will be used.
\item cluster: Resolves futures asynchronously (in parallel) in separate R sessions running typically on one or more machines.
\item remote: Resolves futures asynchronously in a separate R session running on a separate machine, typically on a different network.
#' }}

\item{remotes}{if not \code{NULL} a vector of remote IP addresses}
}
\value{
invisible
}
\description{
Setup a future method
}
\examples{
future_method("multicore")
}
