% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{ntile_bin}
\alias{ntile_bin}
\title{Create ntil bins of a numeric variable}
\usage{
ntile_bin(data, variable = NULL, ntiles = NULL)
}
\arguments{
\item{ntiles}{}
}
\value{

}
\description{
Create ntil bins of a numeric variable
}
\examples{
library(ggplot2)
library(asbtools)

ntile_bin(data = diamonds, "price", 12)
}
