% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{pq_read}
\alias{pq_read}
\title{Read a parquet files}
\usage{
pq_read(
  x = NULL,
  file_path = NULL,
  as_data_frame = T,
  assign_schema = T,
  schema_name = NULL,
  to_duck = F,
  snake_names = F,
  remove_columns = NULL,
  properties = ParquetArrowReaderProperties$create()
)
}
\arguments{
\item{schema_name}{}
}
\value{

}
\description{
Read a parquet files
}
\examples{
library(asbtools)
x = "Desktop/data/usa_spending/fpds/1978.gz.parquet"
pq_read(x = x, to_duck = F)
pq_read(x = x, to_duck = T)
pq_read(x = x, to_duck = F, as_data_frame = F)
pq_read(x = x, to_duck = T, as_data_frame = F)
}
