% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{pq_read_files}
\alias{pq_read_files}
\title{Read set of parquet fules}
\usage{
pq_read_files(
  path = NULL,
  as_data_frame = F,
  to_duck = F,
  exclude_files = NULL,
  schema_file = NULL,
  schema_name = NULL,
  assign_schema = T,
  partitioning = NULL,
  unify_schemas = NULL,
  snake_names = T,
  add_file_name = F,
  to_arrow_table = F,
  return_message = T
)
}
\arguments{
\item{assign_schema}{}
}
\value{

}
\description{
Read set of parquet fules
}
\examples{
library(asbtools)
library(tidyverse)

tbl <- pq_read_files(path = "Desktop/data/usa_spending/contract_archives/solicitations/", as_data_frame = F)

tbl \%>\% count(year_data, sort= T) \%>\% collect()

pq_read_files(path = "Desktop/data/usa_spending/contract_archives/solicitations/", as_data_frame = F)


}
