% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{pq_write}
\alias{pq_write}
\title{Write Arrow Parquet Files}
\usage{
pq_write(
  data = NULL,
  file_path = NULL,
  folder = NULL,
  file_name = NULL,
  chunk_size = NULL,
  version = NULL,
  use_compression = T,
  compression = default_parquet_compression(),
  compression_level = 9,
  use_dictionary = NULL,
  write_statistics = NULL,
  data_page_size = NULL,
  use_deprecated_int96_timestamps = FALSE,
  coerce_timestamps = NULL,
  allow_truncated_timestamps = FALSE,
  properties = NULL,
  arrow_properties = NULL,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Write Arrow Parquet Files
}
