% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{rda_to_pq}
\alias{rda_to_pq}
\title{RDA to Parquet}
\usage{
rda_to_pq(
  rda_file = "Desktop/data/usa_spending/fpds_atom/2018_final.rda",
  unique_data = F,
  sort_column = NULL,
  pq_path = "Desktop/data/",
  pq_folder = "fpds",
  pq_file = NULL,
  use_compression = T,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
RDA to Parquet
}
