% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{read_rda_files}
\alias{read_rda_files}
\title{Read RDA files}
\usage{
read_rda_files(
  path = "Desktop/abresler.github.io/r_packages/govtrackR/data/nsf/nsf_grants/",
  snake_names = F,
  to_arrow_table = T
)
}
\arguments{
\item{path}{a file pah}

\item{to_arrow_table}{}
}
\value{
a \code{tibble}
}
\description{
Read RDA files
}
