% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{read_tsv_gz_files}
\alias{read_tsv_gz_files}
\title{Read TSV}
\usage{
read_tsv_gz_files(path = NULL, snake_names = T, to_arrow_table = F)
}
\arguments{
\item{to_arrow_table}{}
}
\value{

}
\description{
Read TSV
}
