% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{snake_arrow_names}
\alias{snake_arrow_names}
\title{Snake Case Arrow Names}
\usage{
snake_arrow_names(data, case = "snake")
}
\arguments{
\item{data}{an arrow table}

\item{case}{case default to `snake`}
}
\value{

}
\description{
Snake Case Arrow Names
}
