% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_allocate}
\alias{tbl_allocate}
\title{Allocate summarized data}
\usage{
tbl_allocate(
  data,
  allocation_variable = NULL,
  split_variables = NULL,
  split_separator = "\\\\|",
  keep_columns = NULL,
  remove_columns = NULL,
  split_data = TRUE,
  summarise_groups = NULL,
  widen_variable = NULL,
  count_variable = "count",
  distinct_variables = NULL,
  amount_variables = NULL,
  mean_variables = NULL,
  median_variables = NULL,
  min_variables = NULL,
  max_variables = NULL,
  first_variables = NULL,
  last_variables = NULL,
  variance_variables = NULL,
  sd_variables = NULL,
  which_max_variables = NULL,
  which_min_variables = NULL,
  remove_top_amount = TRUE,
  coalesce_numeric = F,
  remove_orginal_allocation = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Allocate summarized data
}
