% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{tbl_arrow}
\alias{tbl_arrow}
\title{To Arrow}
\usage{
tbl_arrow(
  data,
  schema = NULL,
  snake_names = F,
  to_duck = F,
  return_message = T,
  assign_schema = T,
  ...
)
}
\arguments{
\item{assign_schema}{}
}
\value{

}
\description{
To Arrow
}
\examples{
library(asbtools)
mtcars \%>\% tbl_arrow()

}
