% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{tbl_arrow_compute_functions}
\alias{tbl_arrow_compute_functions}
\title{Arrow Compute Functions}
\usage{
tbl_arrow_compute_functions()
}
\value{

}
\description{
Arrow Compute Functions
}
\examples{
tbl_arrow_compute_functions()
}
