% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{tbl_arrow_top_n_groups}
\alias{tbl_arrow_top_n_groups}
\title{Top Arrow Groups}
\usage{
tbl_arrow_top_n_groups(
  data,
  group_variables = NULL,
  top_variables = NULL,
  calculation_variable = NULL,
  filters = NULL,
  top = 1,
  remove_top_amount = T
)
}
\arguments{
\item{remove_top_amount}{}
}
\value{

}
\description{
Top Arrow Groups
}
