% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_assign_dictionary_names}
\alias{tbl_assign_dictionary_names}
\title{Assign a set of names to a tibble}
\usage{
tbl_assign_dictionary_names(data, dictionary_names = NULL, snake_names = F)
}
\arguments{
\item{dictionary_names}{}
}
\value{

}
\description{
Requires a `feature` and `actual` column or
a `tibble` with 2 columns and the `actual` as the second colun
}
\examples{
library(tibble)
dict_iris <- tibble(feature = names(iris), actual = c("SL", "SW", "PL", "PW", "Name of Species"))

tbl_assign_dictionary_names(iris, dict_iris)
tbl_assign_dictionary_names(iris, dict_iris, snake_names = T)

}
