% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{tbl_binr}
\alias{tbl_binr}
\title{Cut Numeric Values Into Evenly Distributed Groups (bins).}
\usage{
tbl_binr(
  data,
  variables = NULL,
  remove_original_variables = F,
  target_bins = NULL,
  max_breaks = NULL,
  exact_groups = F,
  verbose = T,
  errthresh = 0.1,
  minpts = NA
)
}
\arguments{
\item{minpts}{}
}
\value{

}
\description{
Cut Numeric Values Into Evenly Distributed Groups (bins).
}
\examples{
library(ggplot2)
library(tidyverse)
library(asbtools)
tbl_binr(data = diamonds, variables = c("price", "table"), target_bins = 5, max_breaks = 1)


}
