% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_cluster_correlations}
\alias{tbl_cluster_correlations}
\title{Correlation Clusters}
\usage{
tbl_cluster_correlations(
  data,
  exclude_cluster_correlations = T,
  methods = c("som", "cmeans", "emc", "hardcl", "hopach", "h2okmeans", "kmeans",
    "ngas", "pam", "pamk", "spec"),
  correlation_method = "pearson",
  correlation_used = "pairwise.complete.obs",
  diagonal = NA,
  remove_columns = NULL,
  include_logical = T,
  include_factors = F,
  character_to_factor = F,
  full_rank = T
)
}
\arguments{
\item{methods}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{correlation_method}{an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{full_rank}{}
}
\value{

}
\description{
Correlation Clusters
}
