% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_count}
\alias{tbl_count}
\title{Build a tibble of counts}
\usage{
tbl_count(
  data,
  matching_variables = c("^name", "^type"),
  weight_columns = c(NA),
  exclude_columns = NULL,
  include_logical = T,
  include_factor = T,
  unite_features = T,
  numeric_columns = NULL,
  scale_results = F,
  include_proportion = T
)
}
\arguments{
\item{scale_results}{}
}
\value{

}
\description{
Build a tibble of counts
}
