% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_cran_packages}
\alias{tbl_cran_packages}
\title{Tibble of CRAN Packages}
\usage{
tbl_cran_packages(normalize_text = T)
}
\arguments{
\item{normalize_text}{if `TRUE` normalizes title descriptoon to upper}
}
\value{

}
\description{
Tibble of CRAN Packages
}
\examples{
library(asbtools)
library(tidyverse)

cran <- tbl_cran_packages()

cran \%>\% sheldon::regex_keyword_match(text_columns = "title", keywords = "markdown", id_columns = "package")
cran \%>\% sheldon::kwic_keyword_match(text_columns = "title", id_columns = "package")

}
