% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{tbl_cut_variables}
\alias{tbl_cut_variables}
\title{Create Cut Bins from a Numeric Variable}
\usage{
tbl_cut_variables(
  data,
  variables = NULL,
  breaks = NULL,
  override_labels = NULL,
  is_ordered_factor = T,
  digit_label = 10,
  include.lowest = T,
  right = T
)
}
\arguments{
\item{variables}{vector of variable names}

\item{breaks}{if a a single number the number of breaks or a numeric vector with the breaks}

\item{right}{}
}
\value{

}
\description{
Create Cut Bins from a Numeric Variable
}
\examples{

library(ggplot2)
library(asbtools)

tbl_cut_variables(data = diamonds, variables = c("price", "table"), breaks = 10)

tbl_cut_variables(data = diamonds, breaks = c(0, 1000, 5000, 15000, 20000), variable = "price", right = F, include.lowest = F) \%>\% count(price_0005_cut_bins_0_to_20000)


}
