% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_dimension_correlations}
\alias{tbl_dimension_correlations}
\title{Dimension Reduced Correlations}
\usage{
tbl_dimension_correlations(
  data,
  exclude_dimensional_correlations = T,
  methods = c("gng", "cur", "h2oae", "h2oglrm", "ica", "isomap", "kpca", "lle", "mds",
    "nmf", "pca", "spca", "svd", "tsne", "umap"),
  correlation_method = "pearson",
  correlation_used = "pairwise.complete.obs",
  diagonal = NA,
  remove_columns = NULL,
  include_logical = T,
  include_factors = F,
  exclude_clusters = F,
  character_to_factor = F,
  full_rank = T
)
}
\arguments{
\item{methods}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{correlation_method}{an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".}

\item{full_rank}{}
}
\value{

}
\description{
Dimension Reduced Correlations
}
