% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_edit}
\alias{tbl_edit}
\title{Edit Data}
\usage{
tbl_edit(data, file_path = NULL, folder = NULL, file_name = NULL)
}
\arguments{
\item{data}{data frame}

\item{file_path}{if not `NULL` a filepath to save the file}

\item{folder}{folder to save the file}

\item{file_name}{file name}
}
\value{

}
\description{
Edit Data
}
\examples{
tbl_edit(data = iris, file_path = "Desktop/test",  file_name = "iris")
tbl_edit(data = iris, file_path = "Desktop",  file_name = "iris")
}
