% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_nest}
\alias{tbl_nest}
\title{Nest a tibble by grouping variables and specified other variables}
\usage{
tbl_nest(
  data,
  grouping_variables = NULL,
  nesting_variables = NULL,
  data_column_name = NULL
)
}
\arguments{
\item{data}{a `tibble`}

\item{grouping_variables}{vector of variables to group by}

\item{nesting_variables}{if not `NULL` vector of other variables to select}

\item{data_column_name}{if not `NULL` new nested data column name}
}
\value{
`tibble`
}
\description{
Nest a tibble by grouping variables and specified other variables
}
\examples{

gapminder::gapminder \%>\%
tbl_nest(grouping_variables = "year",
data_column_name = "ttdata")


gapminder::gapminder \%>\%
tbl_nest(grouping_variables = c("continent", "country"),
nesting_variables = c("pop", "year"))


}
