% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{tbl_ntile_bin}
\alias{tbl_ntile_bin}
\title{Create ntile bins}
\usage{
tbl_ntile_bin(
  data,
  variables = NULL,
  ntiles = NULL,
  remove_original_bin = F,
  remove_summary_features = T
)
}
\arguments{
\item{variables}{name of the bin variables}

\item{ntiles}{number of ntiles}

\item{remove_original_bin}{remove the orignial numeric feature}

\item{remove_summary_features}{remove summaray fatures}
}
\value{

}
\description{
Create ntile bins
}
\examples{
library(ggplot2)
library(asbtools)

tbl_ntile_bin(diamonds, variables = c("price", "carat"), ntiles = c(5,10))
tbl_ntile_bin(diamonds, variables = c("price"), ntiles = c(5,10))
tbl_ntile_bin(diamonds, variables = c("carat"), ntiles = c(10))

}
