% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{tbl_partition}
\alias{tbl_partition}
\title{Partition tbl}
\usage{
tbl_partition(data, partitions = 10, return_message = T)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Partition tbl
}
