% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_summarise}
\alias{tbl_summarise}
\title{Summarise across tibble}
\usage{
tbl_summarise(
  data,
  group_variables = NULL,
  widen_variable = NULL,
  count_variable = "count",
  distinct_variables = NULL,
  amount_variables = NULL,
  mean_variables = NULL,
  top_variables = NULL,
  calculation_variable = NULL,
  median_variables = NULL,
  min_variables = NULL,
  max_variables = NULL,
  which_max_variables = NULL,
  which_min_variables = NULL,
  unique_variables = NULL,
  unique_concatenator = " | ",
  first_variables = NULL,
  last_variables = NULL,
  variance_variables = NULL,
  sd_variables = NULL,
  coalesce_numeric = F,
  remove_top_amount = T,
  filters = c("UNNAMED", "UNKNOWN"),
  ...
)
}
\arguments{
\item{unique_concatenator}{}
}
\value{

}
\description{
Summarise across tibble
}
\examples{
library(tidyverse)
library(asbviz)
ggplot2::diamonds \%>\%
mutate_if(is.factor, as.character) \%>\%
tbl_summarise(
unique_variables = "clarity",
calculation_variable = "price",
amount_variable = "price"
)

ggplot2::diamonds \%>\%
tbl_summarise(
group = "cut",
unique_variables = "clarity",
calculation_variable = "price",
amount_variable = "price"
)

ggplot2::diamonds \%>\%
tbl_summarise(
widen_variable = "cut",
unique_variables = "clarity",
calculation_variable = "price",
amount_variable = "price"
)



}
