% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tbl_tools.R
\name{tbl_top_n_groups}
\alias{tbl_top_n_groups}
\title{Top Groups}
\usage{
tbl_top_n_groups(
  data,
  group_variables = NULL,
  top_variables = NULL,
  calculation_variable = "amount_award_allocated",
  filters = c("UNNAMED", "UNKNOWN"),
  top = 1,
  remove_top_amount = T
)
}
\arguments{
\item{remove_top_amount}{}
}
\value{

}
\description{
Top Groups
}
