% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{tree_path}
\alias{tree_path}
\title{Return Tree Path}
\usage{
tree_path(path = NULL)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Return Tree Path
}
\examples{

tree_path(path = "Desktop/data/usa_spending/fpds/")

}
