% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{write_partitioned_rda}
\alias{write_partitioned_rda}
\title{Write partioned RDA files}
\usage{
write_partitioned_rda(
  data = NULL,
  partitions = 10,
  include_name = T,
  file_name = NULL,
  folder_path = NULL,
  return_message = T
)
}
\arguments{
\item{partitions}{number of partitions - defaults to 10}

\item{return_message}{}
}
\value{

}
\description{
Write partioned RDA files
}
