% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{write_rda}
\alias{write_rda}
\title{Write single RDA file}
\usage{
write_rda(
  data = NULL,
  file_path = NULL,
  folder = NULL,
  file_name = NULL,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Write single RDA file
}
