% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{write_rda_group_files}
\alias{write_rda_group_files}
\title{Title}
\usage{
write_rda_group_files(
  data = NULL,
  group = NULL,
  include_name = T,
  file_name = NULL,
  folder_path = NULL,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Title
}
