% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{write_tsv_gz}
\alias{write_tsv_gz}
\title{Write TSV.gz file}
\usage{
write_tsv_gz(
  data = NULL,
  file_path = NULL,
  folder = NULL,
  file_name = NULL,
  return_message = T
)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Write TSV.gz file
}
