% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tools.R
\name{pq_read_files}
\alias{pq_read_files}
\title{Read Parquet Files}
\usage{
pq_read_files(
  path = NULL,
  exclude_files = NULL,
  add_file_name = F,
  as_data_frame = T,
  return_message = T
)
}
\arguments{
\item{path}{folder path}

\item{exclude_files}{vector of file names to exclude}

\item{add_file_name}{if `TRUE` appends file name}

\item{return_message}{if `TRUE` returns a message}
}
\value{

}
\description{
Read Parquet Files
}
