% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{tbl_arrow_summarise}
\alias{tbl_arrow_summarise}
\title{Summarise Arrow Table}
\usage{
tbl_arrow_summarise(
  data,
  group_variables = NULL,
  widen_variable = NULL,
  count_variable = "count",
  distinct_variables = NULL,
  amount_variables = NULL,
  mean_variables = NULL,
  top_variables = NULL,
  calculation_variable = NULL,
  median_variables = NULL,
  min_variables = NULL,
  max_variables = NULL,
  which_max_variables = NULL,
  which_min_variables = NULL,
  unique_variables = NULL,
  unique_concatenator = " | ",
  first_variables = NULL,
  last_variables = NULL,
  variance_variables = NULL,
  sd_variables = NULL,
  coalesce_numeric = T,
  remove_top_amount = T,
  filters = NULL,
  to_arrow_table = F,
  ...
)
}
\arguments{
\item{to_arrow_table}{}
}
\value{

}
\description{
Summarise Arrow Table
}
